Update php.md
When installing php74, the default on centos is php54, which is the file /usr/bin/php. The procedure is to copy the php file to php54, delete the php file and replace with a symbolic link to php74 that is named php.  This way you can automatically control which version of PHP you want as default on your system.
There is a "fix" that says to use scl to create a temporary default for your session to run php74, but this method eliminates the hassles, UNLESS you really need to use different verisions of php on your system.